--- a/src/search.cpp
+++ b/src/search.cpp
@@ -30,18 +30,18 @@ namespace search {
       std::fill(back_pointers.get(), back_pointers.get() + rows * cols, '.');
     }
 
-    float *Dynamic::get_score(size_t r, size_t c) {
+    float &Dynamic::get_score(size_t r, size_t c) {
       if ((r > rows - 1) || (c > cols - 1))
         throw std::runtime_error("invalid cost scores access");
 
-      return &scores[r * cols + c];
+      return scores[(r % 2) * cols + c];
     }
 
-    char *Dynamic::get_backpointer(size_t r, size_t c) {
+    char &Dynamic::get_backpointer(size_t r, size_t c) {
       if ((r > rows - 1) || (c > cols - 1))
         throw std::runtime_error("invalid cost back_pointers access");
 
-      return &back_pointers[r * cols + c];
+      return back_pointers[r * cols + c];
     }
 
     void Dynamic::process(std::vector<utils::scoremap> &smap_list) {
@@ -62,10 +62,10 @@ namespace search {
 
       for (size_t r = 0; r < smap_list.size(); ++r) {
         for (size_t c = 0; c < cols - 1; ++c) {
-          best_score = *get_score(r, c + 1);
+          best_score = get_score(r, c + 1);
           pointer = '^';
 
-          score = *get_score(r + 1, c);
+          score = get_score(r + 1, c);
           if (score > best_score) {
             best_score = score;
             pointer = '<';
@@ -73,7 +73,7 @@ namespace search {
 
           boost::unordered_map<utils::sizet_pair, float>::const_iterator got = alignments.find({r, c});
           if (got != alignments.end()) {
-            score = got->second + *get_score(r, c);
+            score = got->second + get_score(r, c);
 
             if (score > best_score) {
               best_score = score;
@@ -81,8 +81,8 @@ namespace search {
             }
           }
 
-          *get_score(r + 1, c + 1) = best_score;
-          *get_backpointer(r, c) = pointer;
+          get_score(r + 1, c + 1) = best_score;
+          get_backpointer(r, c) = pointer;
 
         }
       }
@@ -100,7 +100,7 @@ namespace search {
 
       for (size_t r = 0; r < rows; ++r) {
         for (size_t c = 0; c < cols; ++c) {
-          std::cout << *(get_backpointer(r, c)) << "\t";
+          std::cout << get_backpointer(r, c) << "\t";
         }
         std::cout << "\n";
       }
@@ -114,7 +114,7 @@ namespace search {
       char pointer;
 
       while (i >= 0 && j >= 0) {
-        pointer = *get_backpointer(i, j);
+        pointer = get_backpointer(i, j);
         if (pointer == '^') {
           i -= 1;
         } else if (pointer == '<') {