--- a/src/search.h
+++ b/src/search.h
@@ -36,7 +36,7 @@ namespace search {
 
         ~Dynamic() {};
 
-        double *get_score(size_t r, size_t c);
+        float *get_score(size_t r, size_t c);
 
         char *get_backpointer(size_t r, size_t c);
 
@@ -52,8 +52,8 @@ namespace search {
         size_t rows = 0;
         size_t cols = 0;
 
-        boost::unordered_map<utils::sizet_pair, double> alignments;
-        std::unique_ptr<double[]> scores;
+        boost::unordered_map<utils::sizet_pair, float> alignments;
+        std::unique_ptr<float[]> scores;
         std::unique_ptr<char[]> back_pointers;
 
     };
@@ -126,9 +126,9 @@ namespace search {
 
 
     void FindMatches(utils::matches_vec &matches, std::vector<utils::scoremap> &scorelist, size_t translated_size,
-                     size_t english_size, double threshold = 0);
+                     size_t english_size, float threshold = 0);
 
-    void FilterMatches(utils::matches_vec &matches, std::vector<utils::scoremap> &scorelist, double threshold = 0);
+    void FilterMatches(utils::matches_vec &matches, std::vector<utils::scoremap> &scorelist, float threshold = 0);
 
 
 }