I've replaced the bound-checking vector's `at()` with using the less safe (but faster) `operator[]` operator. Most of these access operations are dictated by ngram_size and iterating from 0 to ngram_size. This eliminated a lot (I mean a lot!) of unnecessary calls to `size()`.

I've also moved the `correct` vector scope outside the hot loop as it was the source of many dynamic allocations but the size never changes & it is completely overwritten in the beginning of the loop anyway.

Without: (compiled for release)
```
time ./bleualign_cpp --bleu-threshold=0.2 ../pt-153-bad.txt > ../pt-153-bad-out~old.txt

________________________________________________________
Executed in  474.10 secs   fish           external
   usr time  444.32 secs  159.00 micros  444.32 secs
   sys time   24.41 secs  1468.00 micros   24.40 secs
```

With:
```
time ./bleualign_cpp --bleu-threshold=0.2 ../pt-153-bad.txt > ../pt-153-bad-out~new.txt

________________________________________________________
Executed in  240.04 secs   fish           external
   usr time  213.35 secs  166.00 micros  213.35 secs
   sys time   22.28 secs  1137.00 micros   22.28 secs
```