@jelmervdl 2020-05-07
Note that his was measured on my laptop (Intel i9, macOS) with llvm, three runs each. I haven't been able to get this performance yet on CSD3…

Also, when compiled for release I see it consume a lot of memory, more than I noticed it using when compiling for debugging (which was running at less than 1/10th of these speeds…)

Let me first figure out what's going on here, as the memory issue is also the source of my frustrations with running it on certain batches on CSD3.