@jelmervdl 2020-05-07
I've found some more low-hanging fruit, changing the peak memory usage from this test case (two documents with about 60,000 sentences each) from 30gb to about 3.4gb. And that also makes it slightly faster again.

```
Executed in  207.45 secs   fish           external
   usr time  204.87 secs  129.00 micros  204.87 secs
   sys time    2.06 secs  706.00 micros    2.06 secs
```

I wonder if something like Hirschberg's algorithm might be a better solution instead? Essentially it's the same sequence alignment problem, right? It's the same complexity, but only needs O(min{n,m}) instead of O(nm) memory.