@jelmervdl 2020-05-07
Sorry, don't merge just yet. I'm investigating issues with CSD3 where it seems to run nice and fast, and then never stops running after printing the output…